link validator with tle client
